Whole overhaul of the conversion script, preparing for converging tools

+++ conversion.py
@@ -1,14 +1,8 @@
 #!/usr/bin/python3
 
-import glob, os, shutil, zipfile, untangle, datetime
+import glob, os, shutil, zipfile, untangle, datetime, getopt
 from sys import argv
 from unrar import rarfile
-
-try:
-    workdir = os.path.abspath(argv[1])
-    os.chdir(workdir)
-except IndexError:
-    workdir = os.getcwd()
 
 def unpackRar(comicFile, dirPath):
     if rarfile.is_rarfile(comicFile):
@@ -33,15 +27,68 @@ def packZip(comicFile, dirPath):
         #Log error here as well?
         return
 
-for file in glob.glob("*.cbr"):
-    dirPath = os.path.splitext(file)[0]
+def main():
+    try:
+        opts, args = getopt.getopt(argv[1:], "h:i:c", ["help", "inputdir=", "convert"])
+    except getopt.GetoptError as err:
+        print(err)
+        exit()
 
-    unpackRar(file, dirPath)
-    packZip(file, dirPath)
+    workDir = None
+    taskConv = False
 
-    if zipfile.is_zipfile(dirPath+".cbz"):
-        shutil.rmtree(os.path.splitext(file)[0])
-        os.remove(file)
-    else:
-        # Error on cleanup
-        return
+    helpText ='''A script for converting CBR files to a CBZ format for openness.
+
+Usage:
+
+-c --convert
+    Grab all CBR files in directory and convert them to CBZ format with checks
+
+-h --help
+    Print this message
+        
+Options:
+
+-i --inputdir=
+    The directory to search for files. If not set, will use current directory.
+'''
+
+    if len(argv) == 1:
+        print("No options. Run with --help for more info.")
+        exit()
+
+    for o, a in opts:
+        if o in ("-h", "--help"):
+            print(helpText)
+            exit()
+        elif o in ("-i", "--inputdir"):
+            workDir = a
+        elif o in ("-c", "--convert"):
+            taskConv = True
+        else:
+            assert False, "unhandled option"
+
+    try:
+        workDir = os.path.abspath(workDir)
+        os.chdir(workDir)
+    except:
+        workDir = os.getcwd()
+
+    # if taskConv, indent rest
+    if taskConv:
+        for file in glob.glob("*.cbr"):
+            dirPath = os.path.splitext(file)[0]
+
+            unpackRar(file, dirPath)
+            packZip(file, dirPath)
+
+            # Integrity check and cleanup
+            if zipfile.is_zipfile(dirPath+".cbz"):
+                shutil.rmtree(os.path.splitext(file)[0])
+                os.remove(file)
+            else:
+                # Error on cleanup
+                return
+        
+if __name__ == '__main__':
+   main()
